Revert "Initial Windows support (#10)" (#11)
This reverts commit 7767da94e87c7559b03eeb41a04218fd46f34468.

This was accidentally force-pushed before merging and included the incorrect changes.
