+++ cmd/package/package.go
@@ -129,10 +129,8 @@ func main() {
 			return packaging.BuildDeb(opt)
 		case "rpm":
 			return packaging.BuildRPM(opt)
-		case "msi":
-			return packaging.BuildMSI(opt)
 		default:
-			return errors.New("type must be one of ('pkg', 'deb', 'rpm', 'msi')")
+			return errors.New("type must be one of ('pkg', 'deb', 'rpm')")
 		}
 	}
